+++ tesseraql-camel-runtime/src/test/java/io/tesseraql/runtime/MessagingRecipeIntegrationTest.java
@@ -152,6 +152,8 @@ class MessagingRecipeIntegrationTest {
                     password: %s
 
                 tesseraql:
+                  app:
+                    name: messaging-recipe
                   outbox:
                     dispatch:
                       fixedDelay: 500ms