+++ tesseraql-camel-runtime/src/test/java/io/tesseraql/runtime/OrgDataScopingIntegrationTest.java
@@ -195,6 +195,8 @@ class OrgDataScopingIntegrationTest {
                   port: 0
 
                 tesseraql:
+                  app:
+                    name: org-data-scoping
                   datasources:
                     main:
                       jdbcUrl: %s