+++ tesseraql-camel-runtime/src/test/java/io/tesseraql/runtime/PollImportSftpIntegrationTest.java
@@ -149,6 +149,8 @@ class PollImportSftpIntegrationTest {
                     password: %s
 
                 tesseraql:
+                  app:
+                    name: poll-import-sftp
                   connectors:
                     poll:
                       allowedHosts: